Add files via upload
logRoutes.js and log.js are supposed to be added under the routes folder.  HoneypotLog,js and HoneypotLogs.js are suppoosed to be added in the models folder. I modified the honeyDeployment.js to meet the needs of the api, so if you guys added the 2 files pls make changes to the honeyDeployment.js as well. LogController.js are supposed to added to the controllers folder.  This elasticsearch.js is supposed to be added to config folder.  Make a few changes to the server.js as well, pls update.